fix: Fix circular import due to make_admin_client and AdminClient being in the same file
